+++ src/rules/no-missing-label-refs.js
@@ -95,8 +95,7 @@ function findMissingReferences(node, sourceCode) {
 // Rule Definition
 //-----------------------------------------------------------------------------
 
-/** @type {NoMissingLabelRefsRuleDefinition} */
-export default {
+export default /** @satisfies {NoMissingLabelRefsRuleDefinition} */ ({
 	meta: {
 		type: "problem",
 		languages: ["markdown/commonmark", "markdown/gfm"],
@@ -181,4 +180,4 @@ export default {
 			},
 		};
 	},
-};
+});